indexes.sql
An index in SQL is a database object that improves the speed of data retrieval on a table by allowing quick searches on one or more columns. While it speeds up query performance, it may slightly slow down insert, update, and delete operations because the index also needs to be updated.